Brings the prototype branch up to date with master, which has moved on
by 841 commits since it was branched.  Three files conflicted:

  src/update.h              the prototype dropped the comment block
                            above the step counters when it added
                            rigidflag; master's comment is restored and
                            rigidflag kept
  doc/Section_commands.txt  master added (k) markers to surf/temp and
                            temp/global/rescale; the fix rigid entry is
                            kept alongside them
  doc/Section_commands.html the same change in the generated table,
                            which reflows across its six columns

Everything else merged cleanly.  The resulting tree is identical to the
prototype commits rebased onto master, which is an independent
derivation of the same result.

+++ src/KOKKOS/fix_vibmode_kokkos.h
@@ -41,7 +41,7 @@ class FixVibmodeKokkos : public FixVibmode {
   void update_custom_kokkos(int, double, double, double, const double *) const;
 
  private:
-  int boltz;
+  double boltz;
 
 #ifndef SPARTA_KOKKOS_EXACT
   Kokkos::Random_XorShift64_Pool<DeviceType> rand_pool;
