+++ src/compute_gas_reaction_grid.cpp
@@ -135,7 +135,7 @@ void ComputeGasReactionGrid::clear()
 {
   cinfo = grid->cinfo;
   if (ncol == 0) memset(vector_grid,0,nglocal*sizeof(double));
-  else if (nglocal) memset(&array_grid[0][0],0,nglocal*ncol*sizeof(double));
+  else if (nglocal) memset(&array_grid[0][0],0,(bigint) nglocal*ncol*sizeof(double));
 }
 
 /* ----------------------------------------------------------------------
@@ -208,7 +208,7 @@ void ComputeGasReactionGrid::reallocate()
   //   would need to implement (un)pack_grid_one() to avoid this
 
   if (ncol == 0) memset(vector_grid,0,nglocal*sizeof(double));
-  else if (nglocal) memset(&array_grid[0][0],0,nglocal*ncol*sizeof(double));
+  else if (nglocal) memset(&array_grid[0][0],0,(bigint) nglocal*ncol*sizeof(double));
 }
 
 /* ----------------------------------------------------------------------
@@ -219,6 +219,6 @@ bigint ComputeGasReactionGrid::memory_usage()
 {
   bigint bytes = 0;
   if (ncol == 0) bytes += nglocal * sizeof(double);    // vector_grid
-  else bytes += nglocal * ncol * sizeof(double);       // array_grid
+  else bytes += (bigint) nglocal * ncol * sizeof(double);  // array_grid
   return bytes;
 }
